+++ examples/get_behaviour/demo.py
@@ -7,7 +7,7 @@ SCHOOL_ID: int = 0
 
 
 def main():
-    server = api.Server(AUTH, USER_ID, SCHOOL_ID)
+    server = api.Client(AUTH, USER_ID, SCHOOL_ID)
     behaviour = (
         server.get_behaviour()
     )  # Get the last 20 (limit default) behaviour points (positive and negative)