+++ examples/timetable/demo.py
@@ -10,7 +10,7 @@ def main():
     # Some notes:
     # The TimetableInterface object is very confusing, even for me, hopefully I will have the time to redesign it
 
-    server = api.Server(AUTH, USER_ID, SCHOOL_ID)
+    server = api.Client(AUTH, USER_ID, SCHOOL_ID)
     timetable = (
         server.get_timetable()
     )  # Get the TimetableInterface object from the start of this week (Monday)